enforce correct class ordering
