+++ open_flamingo/eval/classification_utils.py
@@ -1001,12 +1001,8 @@ IMAGENET_CLASSNAMES = [
     "corn cob",
     "toilet paper",
 ]
-IMAGENET_1K_CLASS_ID_TO_LABEL = dict(
-    zip(range(len(IMAGENET_CLASSNAMES)), IMAGENET_CLASSNAMES)
-)
 
 HM_CLASSNAMES = [
     "no",
     "yes",
 ]
-HM_CLASS_ID_TO_LABEL = {0: "no", 1: "yes"}
