+++ open_flamingo/eval/eval_datasets.py
@@ -5,7 +5,7 @@ from PIL import Image
 from torch.utils.data import Dataset
 from torchvision.datasets import ImageFolder
 
-from open_flamingo.eval.classification_utils import IMAGENET_1K_CLASS_ID_TO_LABEL
+from open_flamingo.eval.classification_utils import IMAGENET_CLASSNAMES
 
 
 class CaptionDataset(Dataset):
@@ -119,10 +119,14 @@ class ImageNetDataset(ImageFolder):
 
     def __init__(self, root, **kwargs):
         super().__init__(root=root, **kwargs)
+        self.class_id_to_name = dict(
+            zip(range(len(IMAGENET_CLASSNAMES)), IMAGENET_CLASSNAMES)
+        )
+
 
     def __getitem__(self, idx):
         sample, target = super().__getitem__(idx)
-        target_label = IMAGENET_1K_CLASS_ID_TO_LABEL[target]
+        target_label = self.class_id_to_name[target]
         return {
             "id": idx,
             "image": sample,
