+++ open_flamingo/eval/evaluate.py
@@ -10,6 +10,7 @@ import numpy as np
 import torch
 from sklearn.metrics import roc_auc_score
 import utils
+import math
 
 from coco_metric import compute_cider, postprocess_captioning_generation
 from eval_datasets import (
@@ -24,9 +25,7 @@ from tqdm import tqdm
 
 from classification_utils import (
     IMAGENET_CLASSNAMES,
-    IMAGENET_1K_CLASS_ID_TO_LABEL,
     HM_CLASSNAMES,
-    HM_CLASS_ID_TO_LABEL,
 )
 
 from eval_model import BaseEvalModel
@@ -1068,7 +1067,6 @@ def evaluate_classification(
         test_dataset = ImageNetDataset(os.path.join(args.imagenet_root, "val"))
         prompt_fn = lambda x: eval_model.get_imagenet_prompt(label=x["class_name"])
         all_class_names = IMAGENET_CLASSNAMES
-        class_id_to_name = IMAGENET_1K_CLASS_ID_TO_LABEL
         k = 5
     elif dataset_name == "hateful_memes":
         train_dataset = HatefulMemesDataset(
@@ -1083,10 +1081,13 @@ def evaluate_classification(
             text=x["ocr"], label=x["class_name"]
         )
         all_class_names = HM_CLASSNAMES
-        class_id_to_name = HM_CLASS_ID_TO_LABEL
         k = 1
     else:
         raise ValueError(f"Unsupported dataset {dataset_name}")
+
+    class_id_to_name = dict(
+        zip(range(len(all_class_names)), all_class_names)
+    )
 
     effective_num_shots = utils.compute_effective_num_shots(num_shots, args.model)
 
@@ -1120,9 +1121,9 @@ def evaluate_classification(
             batch_demo_samples = utils.sample_batch_demos_from_query_set(
                 query_set, effective_num_shots, len(batch["image"])
             )
-
+        
         # set up prompt ensembling
-        num_permutations = 6 if use_prompt_ensembling else 1
+        num_permutations = min(6, math.factorial(effective_num_shots)) if use_prompt_ensembling else 1
         logprobs = []
         for _ in range(num_permutations):
 
@@ -1169,7 +1170,7 @@ def evaluate_classification(
         # compute accuracy
         for i, topk in enumerate(predicted_classnames):
             y_i = batch["class_name"][i]
-            score = torch.exp(predicted_logprobs[i][0] - torch.logsumexp(logprobs[i], dim=0))
+            score = torch.exp(predicted_logprobs[i][0] - torch.logsumexp(logprobs[i], dim=0)).item()
             predictions.append(
                 {
                     "id": batch["id"][i],
